Minor changes and added some cli commands

+++ basecli/validator.py
@@ -12,6 +12,7 @@ from rich.prompt import Prompt
 
 console = Console()
 
+
 def validate(config_path: str):
     if not os.path.exists(config_path):
         os.makedirs(config_path)
@@ -19,15 +20,18 @@ def validate(config_path: str):
     if not os.path.exists(config_path):
         with open(config_path, "w") as f:
             # This version is different from the package version
-            json.dump({"version" : "v1", "default_project_key": " ", "projects": {}}, f)
+            json.dump({"version": "v1", "default_project_key": " ", "default_base": [], "projects": {}}, f)
     try:
         with open(config_path, "r") as f:
             config = json.load(f)
     except json.decoder.JSONDecodeError:
-        console.print("Invalid config file!", style="bold red")
-        console.print("Deleting config file...", style="bold red")
-        os.remove(config_path)
-        sys.exit(1)
+        console.print("JSONDecodeError! Deleting the config file", style="bold red")
+        console.print(f"Config Path: {config_path}", style="bold red")
+        cont = console.input('Please save your config file if you want to keep it. Press Enter to continue...')
+        if cont == "":
+            os.remove(config_path)
+            console.print("Deleted config file!", style="bold green")
+            sys.exit(1)
 
     try:
         if config["version"] != "v1":
@@ -47,6 +51,6 @@ def validate(config_path: str):
         sys.exit(1)
 
     if config["default_project_key"] == " ":
-        return (config_path,False)
+        return (config_path, False)
     elif config["default_project_key"] != " ":
         return (config_path, True)
